backend/requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pinecone>=2.2.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
pillow>=10.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pydantic>=2.4.2
langchain>=0.0.300
requests>=2.31.0
torch
transformers
scikit-learn
nlpaug
numpy
nltk 
# New dependencies for Uber Eats dataset processing
pandas>=2.0.0
kaggle>=1.5.16
pyarrow>=14.0.0
neo4j>=5.14.0  # Optional: If you want to use Neo4j for knowledge graph
spacy>=3.7.2 
